requirements.txt
flask==3.0.0
requests==2.31.0
xhtml2pdf==0.2.11
python-dotenv==1.0.0
werkzeug==3.0.1
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
